code revisions, broke down nav phone states into seperate components, removed any type from context
